[PRMT-4566] Rename tests files
